some warnings  no more sql when local

+++ src/gui/CommandeDialog.java
@@ -292,12 +292,13 @@ public class CommandeDialog extends JDialog implements ActionListener, ListSelec
         } else if (e.getSource() == btn_delUser) {
             Utils.logStream.Log("User " + l_clients.getSelectedValue().getId() + "removed");
             Utils.clients.removeElement(l_clients.getSelectedValue());
-            try{
-                Utils.SQLupdate("DELETE FROM `clients` WHERE `clients`.`id-cli` = \""+ l_clients.getSelectedValue().getId() +"\"");
-            }
-            catch (SQLException ex){
-                Utils.logStream.Error(ex);
-            }
+            if(!Utils.settings.isLocal)
+                try{
+                    Utils.SQLupdate("DELETE FROM `clients` WHERE `clients`.`id-cli` = \""+ l_clients.getSelectedValue().getId() +"\"");
+                }
+                catch (SQLException ex){
+                    Utils.logStream.Error(ex);
+                }
         } else if (e.getSource() == btn_infoUser) {
             new UserInfo(this, l_clients.getSelectedValue(), Utils.commandes).setVisible(true);
         } else if (e.getSource() == btn_prodComm) {
@@ -335,12 +336,13 @@ public class CommandeDialog extends JDialog implements ActionListener, ListSelec
                 Utils.produits.setProdStock();
                 dispose();
             } else { // Edit de commande
-                try{
-                    Utils.SQLupdate(String.format("UPDATE `commandes` SET `id-cli` = \"%s\", `dateCreation` = \"%s\" WHERE `commandes`.`id-com` = \"%s\"", l_clients.getSelectedValue().getId(), dateCreation.get(Calendar.YEAR) + "/" + (dateCreation.get(Calendar.MONTH) + 1) + "/" +dateCreation.get(Calendar.DAY_OF_MONTH), commande.getId()));
-                }
-                catch (SQLException ex){
-                    Utils.logStream.Error(ex);
-                }
+                if(!Utils.settings.isLocal)
+                    try{
+                        Utils.SQLupdate(String.format("UPDATE `commandes` SET `id-cli` = \"%s\", `dateCreation` = \"%s\" WHERE `commandes`.`id-com` = \"%s\"", l_clients.getSelectedValue().getId(), dateCreation.get(Calendar.YEAR) + "/" + (dateCreation.get(Calendar.MONTH) + 1) + "/" +dateCreation.get(Calendar.DAY_OF_MONTH), commande.getId()));
+                    }
+                    catch (SQLException ex){
+                        Utils.logStream.Error(ex);
+                    }
                 commande.setClient(l_clients.getSelectedValue());
                 commande.setDateCreation(dateCreation);
                 commande.emptyEmprunts();
